Fix product + Add UserServices + Add UserRoleRelation + Login + Signup

+++ Ecommerce.Application/Services/RoleService.cs
@@ -1,44 +1,100 @@
-﻿using Ecommerce.Core.Entities;
+﻿using Ecommerce.Application.DTOs;
+using Ecommerce.Application.Interfaces;
+using Ecommerce.Core.Entities;
 using Ecommerce.Core.Interfaces;
 
 namespace Ecommerce.Application.Services;
 
-public class RoleService
+public class RoleService : IRoleServices
 {
     private readonly IRoleRepository _roleRepository;
+
 
     public RoleService(IRoleRepository roleRepository)
     {
         _roleRepository = roleRepository;
     }
-
-    public async Task<Role> GetRoleByIdAsync(Guid id)
+    public  async Task<RoleDto> GetRoleByIdAsync(Guid id)
     {
-        return await _roleRepository.GetRoleByIdAsync(id);
+        var role = await _roleRepository.GetRoleByIdAsync(id);
+        
+        return new RoleDto
+        {
+           Name = role.Name,
+           Id = role.Id,
+           UserRoles = role.UserRoles
+        };
+
+
     }
 
-    public async Task<IEnumerable<Role>> GetAllRolesAsync()
+    public async Task<RoleDto> AddRoleAsync(AddUpdateRoleDto roleDto)
     {
-        return await _roleRepository.GetAllRulesAsync();
-    }
-
-    public async Task<IEnumerable<Role>> GetRoleByName(String name)
-    {
-        return await _roleRepository.GetRolesByName(name);
-    }
-
-    public async Task AddRoleAsync(Role role)
-    {
+        var role = new Role
+        {
+            Id = Guid.NewGuid(),
+            Name = roleDto.Name,
+        };
         await _roleRepository.AddRoleAsync(role);
+
+        return new RoleDto
+        {
+            Name = role.Name,
+            Id = role.Id,
+            UserRoles = role.UserRoles
+        };
+
     }
 
-    public async Task DeleteRoleByIdAsync(Guid id)
+    public async Task<RoleDto> UpdateRoleAsync(Guid id, AddUpdateRoleDto updateRoleDto)
     {
+        var targetRole = await _roleRepository.GetRoleByIdAsync(id);
+
+        targetRole.Name = updateRoleDto.Name;
+        // targetRole.UserRoles = updateRoleDto.UserRoles;
+
+        await _roleRepository.UpdateRoleAsync(targetRole);
+
+        return new RoleDto
+        {
+            Name = targetRole.Name,
+            UserRoles = targetRole.UserRoles,
+            Id = targetRole.Id,
+        };
+    }
+
+    public async Task<bool> DeleteRoleByIdAsync(Guid id)
+    {
+        var targetRole = await _roleRepository.GetRoleByIdAsync(id);
+        if (targetRole == null)
+        {
+            return false;
+        }
+
         await _roleRepository.DeleteRoleByIdAsync(id);
+        return true;
     }
 
-    public async Task UpdateRoleAsync(Role newRole)
+    public async Task<IEnumerable<RoleDto>> GetAllRolesAsync()
     {
-        await _roleRepository.UpdateRoleAsync(newRole);
+        var roles = await _roleRepository.GetAllRulesAsync();
+        return roles.Select(role => new RoleDto
+        {
+            Id = role.Id,
+            Name = role.Name,
+            UserRoles = role.UserRoles
+        });
+    }
+
+    public async Task<IEnumerable<RoleDto>> GetAllRolesByNameAsync(string name)
+    {
+        var roleByName = await _roleRepository.GetRolesByName(name);
+
+        return roleByName.Select(role => new RoleDto
+        {
+            Id = role.Id,
+            Name = role.Name,
+            UserRoles = role.UserRoles,
+        });
     }
 }
